+++ src/hooks/use-compared-ingredients.ts
@@ -2,15 +2,16 @@
 
 import { ReadonlyURLSearchParams } from 'next/navigation';
 import { DetailedDish } from '@/src/types/api.types';
+import pantryIngredients from '@/public/data/ingredients/pantry.json';
 
-export function useComparedIngredients(searchParams: ReadonlyURLSearchParams, meal: DetailedDish): boolean[] | null {
+export function useComparedIngredients(searchParams: ReadonlyURLSearchParams, dish: DetailedDish): boolean[] | null {
     const sourceUrl = searchParams.get('sourceUrl');
 
     if (!sourceUrl) {
         return null;
     }
 
-    const ingredients = sourceUrl.substring(13).split(',');
+    const ingredients = [...sourceUrl.substring(13).split(','), ...pantryIngredients];
 
-    return meal.ingredients.map(ingredient => ingredients.includes(ingredient.name));
+    return dish.ingredients.map(ingredient => ingredients.includes(ingredient.name));
 }